Condense resourcebanner
